Create rank.cpp
This program solves a Permutation & Combination problem of finding rank of a word if all the words, that can be formed by the letters used in the word, are arranged as in a dictionary. I have used a simple mathematical shortcut/algorithm which makes it easy to solve such problems manually!